+++ packages/react/src/components/List/index.ts
@@ -1,4 +1,4 @@
 export { List } from './List';
 export { ListItem } from './ListItem';
-export type { ListProps, ListBorderStyle } from './List';
+export type { ListProps } from './List';
 export type { ListItemProps } from './ListItem';