+++ packages/react/src/components/Typography/Heading/Heading.stories.tsx
@@ -1,0 +1,20 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import { Heading } from './';
+
+const meta: Meta<typeof Heading> = {
+  title: 'Kjernekomponenter/Typography/Heading',
+  component: Heading,
+};
+
+export default meta;
+
+type Story = StoryObj<typeof Heading>;
+
+export const Preview: Story = {
+  args: {
+    children: 'Tittel tekst',
+    spacing: false,
+    size: 'xlarge',
+  },
+};